requirements.txt
flask==3.0.0
flask-socketio==5.3.6
websocket-client==1.6.4
pandas==2.1.3
numpy==1.26.2
eventlet==0.33.3